11/11/2016 auxlab 1.12
1. Last version had a bug in CSignal::timeshift() --while trimming, it
was supposed to take N (trimmed samples) from the middle but it took
from the 0-point.
2. After 1 was sorted out, it kept crashing--the culprit was
datachunk::UpdateBuffer().. memset() was trying to feed the data in an
invalid memory block. Why? Because the block was not allocated because
it was done with BufSize, which I don't use any more. So, BufSize was
cleaned up and memory alloc was done neatly.