Logical data type & Logical operations
• logical statements revised/debugged
* • (sigproc internal) 1) name change (datachunk to body), 2) datatype
logical—this is possible for both audio and non-audio (a logical array
is the result of a logical operation of a non-logical array, so it can
be either audio or non-audio), 3) bufBlockSize is the number of bytes of
each data point (i.e., 8 for double, 16 for complex, 1 for string and
logical), 4) audio_var(logical array) is a cascade audio array (the
signal is non-null only if the condition is met), nonaudio_var(logical)
is an extracted array with the indices meeting the condition (the same
as MATLAB). 5) New grammar: var(scalar) = array means insert array at
scalar-th position for non-audio, scalar-ms for audio. The following
means replace: audiovar(t1~t2)=signal, audiovar(id1:id2)=signal,
nonaudio(id1:id2)=vector, duration of left and right hand sides can be
different (meaning that if so the duration would change after the
statement).